ui/Django/requirements.txt
Django     ==                        3.2.5
django-heroku  ==                    0.3.1
djangorestframework ==                3.13.1
dj-database-url ==                    0.5.0
psycopg2 == 2.8.6
gunicorn == 20.1.0
whitenoise == 6.0.0
drf-yasg          ==                 1.20.0
numpy == 1.20.3
scikit-fuzzy == 0.4.2